# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ai-assistant",
    "description": "企业级开源 AI 助手平台 — RAG + Agent 编排 + MCP 协议",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat",
        "description": "发起一次非流式对话。",
        "operationId": "chat_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Chat Stream",
        "description": "发起流式对话，以 SSE 增量返回管线阶段与最终回复。",
        "operationId": "chat_stream_api_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "以 text/event-stream 持续返回对话事件",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/chat/conversations": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Conversations",
        "description": "列出当前用户可见的会话。",
        "operationId": "list_conversations_api_chat_conversations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationOut"
                  },
                  "type": "array",
                  "title": "Response List Conversations Api Chat Conversations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/chat/conversations/{conversation_id}": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Conversation",
        "description": "获取会话详情（含消息列表）。",
        "operationId": "get_conversation_api_chat_conversations__conversation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "chat"
        ],
        "summary": "Rename Conversation",
        "description": "修改自己的会话标题。当前租户的系统管理员也可修改该租户内的会话。",
        "operationId": "rename_conversation_api_chat_conversations__conversation_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationRename"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "chat"
        ],
        "summary": "Delete Conversation",
        "description": "删除会话及其消息。",
        "operationId": "delete_conversation_api_chat_conversations__conversation_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Delete Conversation Api Chat Conversations  Conversation Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/tools": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Tools",
        "description": "列出当前可用的工具（内置 + MCP）：名称、描述与参数 Schema。\n\n工具清单会暴露服务端可用能力，因此要求认证与 ``agents:read`` 权限，\n不对外公开。",
        "operationId": "list_tools_api_chat_tools_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Tools Api Chat Tools Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "description": "对话请求体。"
      },
      "ChatResponse": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "reply": {
            "type": "string",
            "title": "Reply"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceOut"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          },
          "code_results": {
            "items": {
              "$ref": "#/components/schemas/CodeResultOut"
            },
            "type": "array",
            "title": "Code Results",
            "default": []
          },
          "skill_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skill Names",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "reply"
        ],
        "title": "ChatResponse",
        "description": "非流式对话响应。"
      },
      "CodeResultOut": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "stdout": {
            "type": "string",
            "title": "Stdout",
            "default": ""
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "CodeResultOut",
        "description": "一次代码执行给界面的结果。不含宿主机路径。"
      },
...
      "ConversationDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageOut"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "user_id",
          "title",
          "created_at",
          "updated_at",
          "messages"
        ],
        "title": "ConversationDetail",
        "description": "会话详情（含消息列表）。"
      },
      "ConversationOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "user_id",
          "title",
          "created_at",
          "updated_at"
        ],
        "title": "ConversationOut",
        "description": "会话概要（不含消息体）。"
      },
      "ConversationRename": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ConversationRename",
        "description": "重命名会话。只接受标题。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceOut"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          },
          "code_results": {
            "items": {
              "$ref": "#/components/schemas/CodeResultOut"
            },
            "type": "array",
            "title": "Code Results",
            "default": []
          },
          "skill_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skill Names",
            "default": []
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "complete"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "content",
          "model",
          "created_at"
        ],
        "title": "MessageOut",
        "description": "消息概要。"
      },
...
      "SourceOut": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "section": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Id"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          }
        },
        "type": "object",
        "required": [
          "filename"
        ],
        "title": "SourceOut",
        "description": "一条检索来源。没有页码或段落时对应字段为空。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}